Criados servicos para Categoria, Transacao e Conta Corrente

+++ Controllers/ContaCorrenteController.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using server.Data;
 using server.Models;
+using server.Utils;
 using Microsoft.AspNetCore.Mvc;
 
 namespace server.Controllers
@@ -12,6 +13,8 @@ namespace server.Controllers
     public class ContaCorrenteController
     {
         private readonly DataContext _context;
+        private TransacaoController _transacaoController;
+        private TransacaoUtil _transacaoUtil;
 
         public ContaCorrenteController(DataContext context) => _context = context;
 
@@ -39,7 +42,7 @@ namespace server.Controllers
             ContaCorrente contaCorrenteOriginal = GetById(contaCorrente.Id);
 
             contaCorrenteOriginal.Nome = contaCorrente.Nome;
-            contaCorrenteOriginal.Saldo = contaCorrente.Saldo;
+            contaCorrenteOriginal.SaldoInicial = contaCorrente.SaldoInicial;
 
             _context.ContasCorrentes.Update(contaCorrenteOriginal);
             _context.SaveChanges();
@@ -50,10 +53,50 @@ namespace server.Controllers
         [Route("delete/{id?}")]
         public ContaCorrente Delete(int id)
         {
+
             ContaCorrente contaCorrente = GetById(id);
-            _context.ContasCorrentes.Remove(contaCorrente);
-            _context.SaveChanges();
+            
+            if(_transacaoUtil.ExisteTransacaoComAContaCorrente(contaCorrente)){
+                contaCorrente.Ativo = false;
+                _context.ContasCorrentes.Update(contaCorrente);
+            } else {
+                _context.ContasCorrentes.Remove(contaCorrente);
+                _context.SaveChanges();
+            }
+
             return contaCorrente;
+        }
+
+        [HttpPost]
+        [Route("saldo")]
+        public double CalcularSaldoTotal() {
+            double saldo = 0;
+
+            List<ContaCorrente> listaContasCorrentes = List();
+
+            listaContasCorrentes.ForEach(delegate(ContaCorrente contaCorrente){
+                if(contaCorrente.Ativo){
+                    saldo += CalcularSaldoContaCorrente(new FiltroPesquisa(contaCorrente));
+                }
+            });
+
+            return saldo;
+        }
+
+        [HttpGet]
+        [Route("saldo/{id?}")]
+        public double CalcularSaldoContaCorrente(FiltroPesquisa filtro) {
+
+            ContaCorrente contaCorrente = GetById(filtro.ContaCorrente.Id);
+            double saldoContaCorrente = contaCorrente.SaldoInicial;
+
+            List<Transacao> listaTransacoes = _transacaoUtil.buscarTransacoesFiltradas(filtro);
+
+            listaTransacoes.ForEach(delegate(Transacao transacao){
+                saldoContaCorrente += transacao.Valor;
+            });
+
+            return saldoContaCorrente;
         }
     }
 }
